refactor(domains): move search-translation to domains/catalog (#232)
